refactor: comment out unused yield and time counter logic in OsvJobManager and CRTable
the commented out yield is due to an optimization. in the beginning this was inserted because the thread could yield during the setup of the wake-call of condvar.cc. in that case it still holds the user_mutex and the io-poller could deadlock (because it could not aquire the lock in the poll callback) and job-exec threads starve. to circumvent this, i yielded before the wait was called. this has bad performance. therefore i removed the yield and check if the io was handled before the wait call (in the unlikely case that the io was polled before the thread finished its condvar setup logic). the automatic yield is in that case done by the mutex itself (see the lfmutex.cc implementation)